ServerLauncher.start() waits for startup tasks to complete
- Extract ControllableProcess interface
- Improve testability of ServerLauncher
