+++ geode-core/src/test/java/org/apache/geode/distributed/ServerLauncherTest.java
@@ -14,20 +14,23 @@
  */
 package org.apache.geode.distributed;
 
+import static org.apache.geode.test.awaitility.GeodeAwaitility.await;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.anyBoolean;
 import static org.mockito.ArgumentMatchers.same;
 import static org.mockito.Mockito.atLeast;
 import static org.mockito.Mockito.eq;
 import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.times;
+import static org.mockito.Mockito.spy;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.verifyZeroInteractions;
 import static org.mockito.Mockito.when;
 
 import java.io.IOException;
 import java.util.Collections;
+import java.util.concurrent.CompletableFuture;
 import java.util.function.Consumer;
 
 import org.junit.Before;
@@ -37,8 +40,9 @@ import org.apache.geode.cache.Cache;
 import org.apache.geode.cache.server.CacheServer;
 import org.apache.geode.distributed.ServerLauncher.Builder;
 import org.apache.geode.distributed.internal.InternalDistributedSystem;
-import org.apache.geode.internal.cache.CacheConfig;
+import org.apache.geode.internal.cache.InternalCache;
 import org.apache.geode.internal.cache.control.InternalResourceManager;
+import org.apache.geode.internal.process.ControllableProcess;
 
 /**
  * Unit tests for {@link ServerLauncher}.
@@ -71,29 +75,6 @@ public class ServerLauncherTest {
     assertThat(ServerLauncherParameters.INSTANCE.getHostnameForClients())
         .isEqualTo("hostName4Clients");
     assertThat(ServerLauncherParameters.INSTANCE.isDisableDefaultServer()).isFalse();
-  }
-
-  @Test
-  public void canBeMocked() throws IOException {
-    ServerLauncher launcher = mock(ServerLauncher.class);
-    Cache cache = mock(Cache.class);
-    CacheConfig cacheConfig = mock(CacheConfig.class);
-
-    when(launcher.getCache()).thenReturn(cache);
-    when(launcher.getCacheConfig()).thenReturn(cacheConfig);
-    when(launcher.getId()).thenReturn("ID");
-    when(launcher.isWaiting(eq(cache))).thenReturn(true);
-    when(launcher.isHelping()).thenReturn(true);
-
-    launcher.startCacheServer(cache, 0L);
-
-    verify(launcher, times(1)).startCacheServer(cache, 0L);
-
-    assertThat(launcher.getCache()).isSameAs(cache);
-    assertThat(launcher.getCacheConfig()).isSameAs(cacheConfig);
-    assertThat(launcher.getId()).isSameAs("ID");
-    assertThat(launcher.isWaiting(cache)).isTrue();
-    assertThat(launcher.isHelping()).isTrue();
   }
 
   @Test
@@ -282,18 +263,24 @@ public class ServerLauncherTest {
 
   @Test
   public void startCacheServerStartsCacheServerWithBuilderValues() throws IOException {
-    Cache cache = mock(Cache.class, "Cache");
+    Cache cache = createCache();
     CacheServer cacheServer = mock(CacheServer.class, "CacheServer");
     when(cache.getCacheServers()).thenReturn(Collections.emptyList());
     when(cache.addCacheServer()).thenReturn(cacheServer);
-    ServerLauncher launcher = new Builder().setServerBindAddress(null).setServerPort(11235)
-        .setMaxThreads(10).setMaxConnections(100).setMaxMessageCount(5).setMessageTimeToLive(10000)
-        .setSocketBufferSize(2048).setHostNameForClients("hostName4Clients").build();
+    ServerLauncher launcher = new Builder()
+        .setServerBindAddress(null)
+        .setServerPort(11235)
+        .setMaxThreads(10)
+        .setMaxConnections(100)
+        .setMaxMessageCount(5)
+        .setMessageTimeToLive(10000)
+        .setSocketBufferSize(2048)
+        .setHostNameForClients("hostName4Clients")
+        .setServerLauncherCacheProvider((a, b) -> cache)
+        .setControllableProcess(mock(ControllableProcess.class))
+        .build();
 
-    final InternalResourceManager internalResourceManager = mock(InternalResourceManager.class);
-    when(cache.getResourceManager()).thenReturn(internalResourceManager);
-
-    launcher.startCacheServer(cache, 0L);
+    launcher.start();
 
     verify(cacheServer).setBindAddress(null);
     verify(cacheServer).setPort(eq(11235));
@@ -308,33 +295,34 @@ public class ServerLauncherTest {
 
   @Test
   public void startCacheServerDoesNothingWhenDefaultServerDisabled() throws IOException {
-    Cache cache = mock(Cache.class, "Cache");
+    Cache cache = createCache();
     CacheServer cacheServer = mock(CacheServer.class, "CacheServer");
     when(cache.getCacheServers()).thenReturn(Collections.emptyList());
     when(cache.addCacheServer()).thenReturn(cacheServer);
-    ServerLauncher launcher = new Builder().setDisableDefaultServer(true).build();
+    ServerLauncher launcher = new Builder()
+        .setDisableDefaultServer(true)
+        .setServerLauncherCacheProvider((a, b) -> cache)
+        .setControllableProcess(mock(ControllableProcess.class))
+        .build();
 
-    final InternalResourceManager internalResourceManager = mock(InternalResourceManager.class);
-    when(cache.getResourceManager()).thenReturn(internalResourceManager);
-
-    launcher.startCacheServer(cache, 0L);
+    launcher.start();
 
     verifyZeroInteractions(cacheServer);
   }
 
   @Test
   public void startCacheServerDoesNothingWhenCacheServerAlreadyExists() throws IOException {
-    Cache cache = mock(Cache.class, "Cache");
+    Cache cache = createCache();
     CacheServer cacheServer1 = mock(CacheServer.class, "CacheServer1");
     CacheServer cacheServer2 = mock(CacheServer.class, "CacheServer2");
     when(cache.getCacheServers()).thenReturn(Collections.singletonList(cacheServer1));
     when(cache.addCacheServer()).thenReturn(cacheServer1);
-    ServerLauncher launcher = new Builder().build();
+    ServerLauncher launcher = new Builder()
+        .setServerLauncherCacheProvider((a, b) -> cache)
+        .setControllableProcess(mock(ControllableProcess.class))
+        .build();
 
-    final InternalResourceManager internalResourceManager = mock(InternalResourceManager.class);
-    when(cache.getResourceManager()).thenReturn(internalResourceManager);
-
-    launcher.startCacheServer(cache, 0L);
+    launcher.start();
 
     verifyZeroInteractions(cacheServer2);
   }
@@ -342,14 +330,15 @@ public class ServerLauncherTest {
   @Test
   public void startCacheServerPassesStartupCompletionActionToResourceManager() throws IOException {
     Runnable startupCompletionAction = mock(Runnable.class);
-    ServerLauncher serverLauncher = new Builder()
-        .setStartupCompletionAction(startupCompletionAction)
-        .build();
-
     InternalResourceManager internalResourceManager = mock(InternalResourceManager.class);
     Cache cache = createCache(internalResourceManager);
+    ServerLauncher serverLauncher = new Builder()
+        .setStartupCompletionAction(startupCompletionAction)
+        .setServerLauncherCacheProvider((a, b) -> cache)
+        .setControllableProcess(mock(ControllableProcess.class))
+        .build();
 
-    serverLauncher.startCacheServer(cache, 0L);
+    serverLauncher.start();
 
     verify(internalResourceManager)
         .runWhenStartupTasksComplete(same(startupCompletionAction), any());
@@ -359,14 +348,15 @@ public class ServerLauncherTest {
   public void startCacheServerPassesStartupExceptionActionToResourceManager() throws IOException {
     @SuppressWarnings("unchecked")
     Consumer<Throwable> startupExceptionAction = mock(Consumer.class);
-    ServerLauncher serverLauncher = new Builder()
-        .setStartupExceptionAction(startupExceptionAction)
-        .build();
-
     InternalResourceManager internalResourceManager = mock(InternalResourceManager.class);
     Cache cache = createCache(internalResourceManager);
+    ServerLauncher serverLauncher = new Builder()
+        .setStartupExceptionAction(startupExceptionAction)
+        .setServerLauncherCacheProvider((a, b) -> cache)
+        .setControllableProcess(mock(ControllableProcess.class))
+        .build();
 
-    serverLauncher.startCacheServer(cache, 0L);
+    serverLauncher.start();
 
     verify(internalResourceManager)
         .runWhenStartupTasksComplete(any(), same(startupExceptionAction));
@@ -381,22 +371,70 @@ public class ServerLauncherTest {
 
     ServerLauncher serverLauncher = new Builder()
         .setServerLauncherCacheProvider(serverLauncherCacheProvider)
+        .setControllableProcess(mock(ControllableProcess.class))
         .build();
     serverLauncher.start();
 
     assertThat(serverLauncher.getCache()).isSameAs(cacheFromBuilder);
   }
 
-  private Cache createCache() {
+  @Test
+  public void startUsesControllableProcessFromBuilder() {
+    ServerLauncherCacheProvider serverLauncherCacheProvider =
+        mock(ServerLauncherCacheProvider.class);
+    ControllableProcess controllableProcess = mock(ControllableProcess.class);
+    Cache cacheFromBuilder = createCache();
+    when(serverLauncherCacheProvider.createCache(any(), any())).thenReturn(cacheFromBuilder);
+
+    ServerLauncher serverLauncher = new Builder()
+        .setServerLauncherCacheProvider(serverLauncherCacheProvider)
+        .setControllableProcess(controllableProcess)
+        .build();
+    serverLauncher.start();
+
+    serverLauncher.stop();
+
+    verify(controllableProcess).stop(anyBoolean());
+  }
+
+  @Test
+  public void startWaitsForStartupTasksToComplete() {
+    ServerLauncherCacheProvider serverLauncherCacheProvider =
+        mock(ServerLauncherCacheProvider.class);
+    InternalResourceManager internalResourceManager = mock(InternalResourceManager.class);
+    InternalCache cacheFromBuilder = createCache(internalResourceManager);
+    when(serverLauncherCacheProvider.createCache(any(), any())).thenReturn(cacheFromBuilder);
+
+    ServerLauncher serverLauncher = new Builder()
+        .setServerLauncherCacheProvider(serverLauncherCacheProvider)
+        .build();
+
+    CompletableFuture<Void> startupTasks = spy(new CompletableFuture<>());
+    when(internalResourceManager.runWhenStartupTasksComplete(any(), any()))
+        .thenReturn(startupTasks);
+
+    CompletableFuture<Void> serverLauncherStart = CompletableFuture.runAsync(serverLauncher::start);
+
+    await().untilAsserted(() -> verify(startupTasks).join());
+    assertThat(serverLauncherStart).isNotDone();
+
+    startupTasks.complete(null);
+
+    await().untilAsserted(() -> assertThat(serverLauncherStart).isDone());
+  }
+
+  private InternalCache createCache() {
     return createCache(mock(InternalResourceManager.class));
   }
 
-  private Cache createCache(InternalResourceManager internalResourceManager) {
-    Cache cache = mock(Cache.class);
+  private InternalCache createCache(InternalResourceManager internalResourceManager) {
+    InternalCache cache = mock(InternalCache.class);
     CacheServer cacheServer = mock(CacheServer.class);
 
     when(cache.addCacheServer()).thenReturn(cacheServer);
     when(cache.getResourceManager()).thenReturn(internalResourceManager);
+    when(internalResourceManager.runWhenStartupTasksComplete(any(), any()))
+        .thenReturn(CompletableFuture.completedFuture(null));
 
     return cache;
   }
